Wired 0.7.1 porting — Phase 1 (wxWidgets 3.x migration) =============================================================
Target host: Ubuntu 24.04 LTS (noble), x86_64, wxWidgets 3.2.4,
autoconf 2.71, automake 1.16.5, libtool 2.4.7, gettext 0.21, gcc 13.

Phase 1 makes the tree COMPILE and LINK against wxGTK 3.2 and produces a
launching `wired` binary (splash -> main window, no crash). All fixes are
the wx2 -> wx3 API breaks audited in AGENTS_PORTING.md, plus the hard
link-stage issues found while actually building.

Changes
-------

wired/src/gui/MainWindow.cpp
  - wxThread::Create() is obsolete in wx3 (Run() creates the thread now);
    dropped the MidiEngine/Seq Create() error blocks, kept Run() and the
    Seq SetPriority() call (still valid pre-Run).
  - Shutdown loop iterated the LIVE wxArrayThread by index while threads
    remove themselves in OnExit() -> index-shift/use-after-free. Snapshot
    the thread pointers into a local wxArrayThread first, then Delete().
    (Delete() itself is fine on running detached threads in wx3.)
  - wxMULTIPLE -> wxFD_MULTIPLE (OnImportWave), wxSAVE -> wxFD_SAVE
    (OnExportWave, OnExportMIDI, OnSaveML).
  - wxGetDiskSpace: left as-is -- wxDiskspaceSize_t is a typedef for
    wxLongLong when wxUSE_LONGLONG, so the existing call compiles.

wired/src/gui/MainApp.{h,cpp}
  - PNG/GIF handler registration needed the real wx3 headers:
    <wx/imagpng.h> / <wx/imaggif.h> (wx/png.h, wx/gif.h don't exist).
  - g_cmdLineDesc: wx3 wxCmdLineEntryDesc uses const char* fields; dropped
    the wxT() wrappers (were breaking on the wxChar* initializer).

wired/src/gui/Wizard.h
  - <wx/listbase.h>/<wx/listctrl.h> were guarded by #ifdef _WIN32 but the
    wxListCtrl member is used on all platforms -> includes unconditional.

wired/src/gui/MLTree.cpp
  - wxTHICK_FRAME removed in wx3 -> wxRESIZE_BORDER.

wired/src/gui/Rack.cpp
  - menu->Append(id, text, false) relied on wx2's public wxString(int)
    ctor (private in wx3); the `false` was dead -- dropped it.

wired/src/gui/Pattern.cpp
  - GetTextExtent(..., long*, long*) -> wxCoord* (int) in wx3; local
    x/y/w/h changed to wxCoord.

wired/src/gui/SettingWindow.cpp
  - wxTR_EXTENDED removed in wx3 -> dropped (extended selection default).

wired/src/libs/WiredWidgets/src/ChoiceButton.cpp
  - wxPen(wxT("BLACK"), ...) string ctor removed -> wxPen(wxColour("BLACK")).

wired/src/libs/WiredWidgets/src/FaderCtrl.cpp
  - WXK_NEXT / WXK_PRIOR removed in wx3 -> WXK_PAGEDOWN / WXK_PAGEUP.

wired/src/plugins/HostCallback.cpp
  - wiredHostVendorName returned (void*)_("P31") -- wx3 _() returns a
    const wxString&, not a pointer; now appends into the caller's wxString
    like the product-name case does.
  - `if (lastOpenDir && ...)` relied on wxString->bool in wx2; now uses
    !lastOpenDir.IsEmpty() (and same for lastSaveDir).
  - wxSAVE -> wxFD_SAVE in the plugin file dialogs.

wired/src/dssi/WiredExternalPluginGui.cpp
  - wxString(wxString, *wxConvCurrent) wide-ctor cluster (~13 sites) ->
    stripped the redundant conv arg (the Descriptor.Name one kept its
    conv since that field is a wxString in this struct, not char*).

wired/src/audio/WaveFile.cpp
  - wxString(sf_strerror(0), wxConvCurrent) passed the wxMBConv POINTER;
    wx3 ctor wants a reference -> *wxConvCurrent.

wired/src/editmidi/Key.cpp and plugins/akai_sampler/{ASKey,ASClavier}.cpp
  - wxWindow::ProcessEvent() became protected in wx3 -> call
    GetEventHandler()->ProcessEvent() instead. Key.cpp printf also now
    uses note.ToUTF8().data() (was casting c_str()).

wired/src/editmidi/{EditMidi,RulerMidi}.cpp,
  plugins/akai_sampler/AkaiSampler.cpp, samplerate/WiredSampleRate.cpp
  - wxTHICK_FRAME -> wxRESIZE_BORDER (remaining sites).

wired/src/plugins/filter/FilterPlug.cpp
  - wxBitmap(wxImage*) has no wx3 ctor -> dereference the pointer.

wired/src/plugins/{reverb,delay,filter,chorus}/...Plug.cpp
  - wxString(STR_*, *wxConvCurrent) wide-ctor cluster (~40 sites) ->
    stripped the redundant conv arg (STR_* is already a wxString).

wired/src/plugins/loop_sampler/LoopSampler.cpp
  - Session save wrote Wave->Filename.c_str() + size() -- wchar bytes in
    wx3; now mb_str() length-prefixed (matches the Load side which reads
    bytes into a char[]).

wired/src/wiredvideo/WiredMplayer.{h,cpp}
  - SendMPlayerMessage took const wxChar* and write()d wide chars to the
    mplayer pipe; now takes const wxString& and writes message.mb_str().

wired/src/wiredvideo/WiredVideo.cpp
  - wxOPEN -> wxFD_OPEN; stripped the two wide-ctor wxString(...,conv).

wired/src/sequencer/Sequencer.cpp
  - wxString(wxT("wired_click.wav"), *wxConvCurrent) wide-ctor -> plain.
  - mPart = mFilename.substr(...).c_str()[0] - 64: wxUniChar - int is
    ambiguous in wx3 -> cast the char to int first.

wired/src/undo/cImportMidiAction.cpp
  - Same wxUniChar - int fix as Sequencer.cpp.

wired/src/codec/WiredCodec.cpp, fileloader/FileConversion.cpp,
  gui/MLTree.cpp
  - Remaining wxString(wxT(...), *wxConvCurrent) wide-ctor sites.

wired/src/Makefile.am  (link stage -- needed for a working binary)
  - `wired` had empty _SOURCES so automake used --tag=CC (gcc); undefined
    std:: symbols. Set wired_LINK = $(CXXLINK) with an explicit CXXLD/
    CXXLINK (--tag=CXX).
  - The -l wx/portaudio/xml2/samplerate/sndfile flags were in wired_LDFLAGS,
    which libtool places BEFORE the static archives in LDADD -> unresolved
    symbols. Moved them to the end of wired_LDADD (and the PORTMIDI block
    now appends -lasound via LDADD too).
  - -Wl,--no-as-needed: libtool hoists the .so DEPS_LIBS before the .a
    archives; without this, the linker drops libWiredWidgets/libWiredAkai
    on --as-needed and their symbols go undefined.

Environment note
----------------
- Requires libwxgtk-media3.2 (wx-config --libs media,core) -- the media
  library must be installed or the link fails with -lwx_gtk3u_media.

Verification
------------
- `make` completes (exit 0) across the whole tree.
- `src/wired` links and launches: stays alive (no segfault) for the full
  test window with a real DISPLAY; no startup error output.
- All remaining compile/link noise is pre-existing warnings (missing
  return statements, VLAs), not errors.

Next (Phase 2)
--------------
Audio engine correctness: channel-count stream open + bounds checks
(AudioEngine.cpp:375,384), Device.cpp latency swap, RingBuffer atomics,
WiredSampleRate data_out/memset/double-free bugs, WaveFile sf_close(NULL).

+++ wired/src/undo/cImportMidiAction.cpp
@@ -169,7 +169,7 @@ cImportAkaiAction::cImportAkaiAction (wxString& path, trackType kind)
   path = path.substr(path.find(wxT(":"), 0) + 1, path.size() - path.find(wxT(":"), 0));
   mPath = path.substr(10, path.size() - 10);
   int pos = mPath.find(wxT("/"), 0);
-  mPart = mPath.substr(0, pos).c_str()[0] - 64;
+  mPart = (int)(mPath.substr(0, pos).c_str()[0]) - 64;
   mPath = mPath.substr(pos, mPath.size() - pos);
   int opos = 0;
   while ((pos = mPath.find(wxT("/"), opos)) != wxString::npos)
